--- js/big-picture.js
@@ -1,1 +1,0 @@
-console.log('Модуль большой фотки подключен');
